// src/app/interfaces/employee.interface.ts
export interface Employee {
  _id: string
  createdAt: string
  email: string
  uid: string
  isActive: boolean
  roles: string[]
  emailVerified: boolean
  __v: number
  dateOfBirth: string
  documentNumber: string
  documentType: string
  lastname: string
  name: string
  commission: string
  identificationNumber: string
  phoneNumber: string
}